# AI文档/OpenAPI.yaml
openapi: 3.0.3
info:
  title: 兜兜有钱 API
  version: 0.1.0
  description: |
    兜兜有钱微信记账小程序接口契约。

    设计契约：
    1. 前后端开发前，优先编写并落盘接口契约文档到本文件。
    2. 所有非登录接口必须明确鉴权方式。
    3. 所有业务接口必须明确权限点。当前一期权限点以接口权限和当前用户数据权限为主，不引入菜单/RBAC 体系。
    4. 用户最终通过微信 openid 识别；后端使用 wxCode 换取 openid，创建或识别 users.openid 后签发 JWT。
servers:
  - url: /api
    description: 默认 API 前缀
...
security:
  - bearerAuth: []
paths:
...
  /tags:
    get:
      tags: [Tags]
      summary: 查询标签列表
      x-permissions:
        auth: bearer
        interface: tags:list
        dataScope: current-user
      responses:
        "200":
          description: 标签列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Tag"
    post:
      tags: [Tags]
      summary: 新增标签
      x-permissions:
        auth: bearer
        interface: tags:create
        dataScope: current-user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagCreateRequest"
      responses:
        "200":
          description: 新增成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Tag"
  /tags/{tagId}:
    patch:
      tags: [Tags]
      summary: 重命名标签
      x-permissions:
        auth: bearer
        interface: tags:update
        dataScope: current-user-owned-resource
      parameters:
        - $ref: "#/components/parameters/TagId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagUpdateRequest"
      responses:
        "200":
          description: 编辑成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Tag"
    delete:
      tags: [Tags]
      summary: 删除标签
      x-permissions:
        auth: bearer
        interface: tags:delete
        dataScope: current-user-owned-resource
      parameters:
        - $ref: "#/components/parameters/TagId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        除微信登录接口外，其他接口均使用 JWT Bearer 鉴权。
        后端从 JWT 中解析当前 userId，并基于 userId 做当前用户数据隔离。
  parameters:
...
    TagId:
      name: tagId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    Ok:
      description: 操作成功
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessEnvelope"
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [code, message, traceId]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          description: 成功响应数据，不同接口返回不同结构；无数据时可为空对象或 null
        traceId:
          type: string
          example: 018f0d8b9d2c4b91
...
    Tag:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    TagCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 20
    TagUpdateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 20
...